fix numpy typing
