+++ test_elasticsearch/test_server/test_vectorstore/conftest.py
@@ -18,20 +18,10 @@
 import uuid
 
 import pytest
-from elastic_transport import Transport
 
 from ...utils import wipe_cluster
 from ..conftest import _create
-
-
-class RequestSavingTransport(Transport):
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self.requests: list = []
-
-    def perform_request(self, *args, **kwargs):
-        self.requests.append(kwargs)
-        return super().perform_request(*args, **kwargs)
+from ._test_utils import RequestSavingTransport
 
 
 @pytest.fixture(scope="function")
